This updates to the latest JRuby that doesn't need asm. Moderm JRuby versions need it, but have version > 4.0, which conflicts with the 3.x version Groovy depends on. This makes it work with Play 1.2.7 again.
