+++ src/play.plugins
@@ -1,1 +1,1 @@
-100:play.modules.sass.Plugin
+100:play.modules.sass.SassPlugin
